refactor(core): extract HTTP error layer to @appstrate/core/api-errors
Move ApiError, factory helpers (notFound, parseBody, forbidden, ...), and the
asRecord JSONB narrower into @appstrate/core/api-errors. apps/api preserves
the old import paths via transparent re-exports — zero call-site changes.

Why: make the HTTP error layer importable from the published core package so
external modules (loaded via MODULES=<npm-pkg>) don't need to reach into
apps/api/src/lib/ for error classes and validation helpers.
