+++ apps/api/src/lib/safe-json.ts
@@ -1,11 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-
-/** Safely narrow a JSONB value to a record, returning {} if null/non-object/array. */
-export function asRecord(val: unknown): Record<string, unknown> {
-  return val !== null && typeof val === "object" && !Array.isArray(val)
-    ? (val as Record<string, unknown>)
-    : {};
-}
+export { asRecord } from "@appstrate/core/api-errors";
 
 /** Safely narrow a JSONB value to a record, returning null if not a plain object. */
 export function asRecordOrNull(val: unknown): Record<string, unknown> | null {
