+++ packages/core/package.json
@@ -42,6 +42,7 @@
     "./logger": "./src/logger.ts",
     "./env": "./src/env.ts",
     "./ajv": "./src/ajv.ts",
+    "./api-errors": "./src/api-errors.ts",
     "./storage": "./src/storage.ts",
     "./storage-s3": "./src/storage-s3.ts",
     "./storage-fs": "./src/storage-fs.ts",
